Change directory from birthregistrationmatching to brm
