+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             uk.gov.hmrc.birthregistrationmatching.controllers.MicroserviceHelloWorld.hello
+GET        /hello-world             uk.gov.hmrc.brm.controllers.MicroserviceHelloWorld.hello
