+++ conf/application.conf
@@ -44,7 +44,7 @@ application.langs="en"
 # ~~~~~
 # Define the Global object class for this application.
 # Default to Global in the root package.
-application.global=uk.gov.hmrc.birthregistrationmatching.config.MicroserviceGlobal
+application.global= uk.gov.hmrc.brm.config.MicroserviceGlobal
 
 # Router
 # ~~~~~
@@ -75,7 +75,7 @@ controllers {
       needsAuditing = false
     }
 
-  uk.gov.hmrc.birthregistrationmatching.controllers.MicroserviceHelloWorld = {
+  uk.gov.hmrc.brm.controllers.MicroserviceHelloWorld = {
       needsAuth = false
       needsLogging = false
       needsAuditing = false
